Version: 0.0.8 Branch: merlyan_transactions Date: 05.07.2017 Author: AlexMerlyan
Working: maintenance
Changes:
Next:

Comment:

+++ src/main/java/projectpackage/controllers/MaintenanceController.java
@@ -1,5 +1,6 @@
 package projectpackage.controllers;
 
+import org.apache.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.hateoas.Resource;
 import org.springframework.http.HttpStatus;
@@ -9,16 +10,18 @@ import org.springframework.web.bind.annotation.*;
 import projectpackage.dto.IUDAnswer;
 import projectpackage.model.auth.User;
 import projectpackage.model.maintenances.Maintenance;
-import projectpackage.service.MessageBook;
+import projectpackage.repository.support.daoexceptions.DeletedObjectNotExistsException;
+import projectpackage.repository.support.daoexceptions.ReferenceBreakException;
+import projectpackage.repository.support.daoexceptions.WrongEntityIdException;
 import projectpackage.service.maintenanceservice.MaintenanceService;
+import projectpackage.service.support.ServiceUtils;
 
 import javax.cache.annotation.CacheRemoveAll;
 import javax.cache.annotation.CacheResult;
 import javax.servlet.http.HttpServletRequest;
 import java.util.List;
 
-import static org.springframework.hateoas.mvc.ControllerLinkBuilder.linkTo;
-import static org.springframework.hateoas.mvc.ControllerLinkBuilder.methodOn;
+import static projectpackage.service.MessageBook.*;
 
 /**
  * Created by Arizel on 28.05.2017.
@@ -27,8 +30,13 @@ import static org.springframework.hateoas.mvc.ControllerLinkBuilder.methodOn;
 @RequestMapping("/maintenances")
 public class MaintenanceController {
 
+    private static final Logger LOGGER = Logger.getLogger(MaintenanceController.class);
+
     @Autowired
     MaintenanceService maintenanceService;
+
+    @Autowired
+    ServiceUtils serviceUtils;
 
     @ResponseStatus(HttpStatus.OK)
     @CacheResult(cacheName = "maintenanceList")
@@ -38,7 +46,7 @@ public class MaintenanceController {
 
     }
 
-    @ResponseStatus(HttpStatus.ACCEPTED)
+    @ResponseStatus(HttpStatus.OK)
     @RequestMapping(value = "/{id}", method = RequestMethod.GET, produces = {MediaType.APPLICATION_JSON_UTF8_VALUE})
     public ResponseEntity<Resource<Maintenance>> getMaintenance(@PathVariable("id") Integer id, HttpServletRequest request) {
         User thisUser = (User) request.getSession().getAttribute("USER");
@@ -47,11 +55,7 @@ public class MaintenanceController {
         Resource<Maintenance> resource = new Resource<>(Maintenance);
         HttpStatus status;
         if (null != Maintenance) {
-            if (thisUser.getRole().getRoleName().equals("ADMIN")) {
-                resource.add(linkTo(methodOn(MaintenanceController.class).deleteMaintenance(Maintenance.getObjectId())).withRel("delete"));
-            }
-            resource.add(linkTo(methodOn(MaintenanceController.class).updateMaintenance(Maintenance.getObjectId(), Maintenance)).withRel("update"));
-            status = HttpStatus.ACCEPTED;
+            status = HttpStatus.OK;
         } else {
             status = HttpStatus.BAD_REQUEST;
         }
@@ -62,36 +66,81 @@ public class MaintenanceController {
 
     @CacheRemoveAll(cacheName = "maintenanceList")
     @RequestMapping(method = RequestMethod.POST, produces = {MediaType.APPLICATION_JSON_UTF8_VALUE}, consumes = {MediaType.APPLICATION_JSON_UTF8_VALUE})
-    public ResponseEntity<IUDAnswer> createMaintenance(@RequestBody Maintenance newMaintenance) {
-        IUDAnswer result = maintenanceService.insertMaintenance(newMaintenance);
+    public ResponseEntity<IUDAnswer> createMaintenance(@RequestBody Maintenance newMaintenance, HttpServletRequest request) {
+        User user = (User) request.getSession().getAttribute("USER");
+        IUDAnswer iudAnswer = serviceUtils.checkSessionAdminAndData(user, newMaintenance);
+        if (!iudAnswer.isSuccessful()) {
+            return new ResponseEntity<IUDAnswer>(iudAnswer, HttpStatus.BAD_REQUEST);
+        }
 
-        HttpStatus status = result.isSuccessful() ? HttpStatus.CREATED : HttpStatus.BAD_REQUEST;
-
-        return new ResponseEntity<IUDAnswer>(result, status);
+        try {
+            iudAnswer = maintenanceService.insertMaintenance(newMaintenance);
+        } catch (IllegalArgumentException e) {
+            LOGGER.warn(WRONG_FIELD, e);
+            return new ResponseEntity<IUDAnswer>(new IUDAnswer(false, WRONG_FIELD, e.getMessage()), HttpStatus.BAD_REQUEST);
+        }
+        HttpStatus status;
+        if (iudAnswer != null && iudAnswer.isSuccessful()) {
+            status = HttpStatus.OK;
+        } else status = HttpStatus.BAD_REQUEST;
+        ResponseEntity<IUDAnswer> responseEntity = new ResponseEntity<IUDAnswer>(iudAnswer, status);
+        return responseEntity;
     }
 
     @CacheRemoveAll(cacheName = "maintenanceList")
     @RequestMapping(value = "/{id}", method = RequestMethod.PUT, produces = {MediaType.APPLICATION_JSON_UTF8_VALUE}, consumes = {MediaType.APPLICATION_JSON_UTF8_VALUE})
-    public ResponseEntity<IUDAnswer> updateMaintenance(@PathVariable("id") Integer id, @RequestBody Maintenance changedMaintenance) {
-        if (id == null) {
-            return new ResponseEntity<IUDAnswer>(new IUDAnswer(id, false, MessageBook.WRONG_UPDATE_ID),
-                    HttpStatus.NOT_ACCEPTABLE);
+    public ResponseEntity<IUDAnswer> updateMaintenance(@PathVariable("id") Integer id,
+                                                       @RequestBody Maintenance changedMaintenance,
+                                                       HttpServletRequest request) {
+        User user = (User) request.getSession().getAttribute("USER");
+        IUDAnswer iudAnswer = serviceUtils.checkSessionAdminAndData(user, changedMaintenance, id);
+        if (!iudAnswer.isSuccessful()) {
+            return new ResponseEntity<IUDAnswer>(iudAnswer, HttpStatus.BAD_REQUEST);
         }
 
-        IUDAnswer result = maintenanceService.updateMaintenance(id, changedMaintenance);
-
-        HttpStatus status = result.isSuccessful() ? HttpStatus.ACCEPTED : HttpStatus.BAD_REQUEST;
-
-        return new ResponseEntity<IUDAnswer>(result, status);
+        try {
+            iudAnswer = maintenanceService.updateMaintenance(id, changedMaintenance);
+        } catch (IllegalArgumentException e) {
+            LOGGER.warn(WRONG_FIELD, e);
+            return new ResponseEntity<IUDAnswer>(new IUDAnswer(id, false, WRONG_FIELD, e.getMessage()), HttpStatus.BAD_REQUEST);
+        }
+        HttpStatus status;
+        if (iudAnswer != null && iudAnswer.isSuccessful()) {
+            status = HttpStatus.OK;
+        } else status = HttpStatus.BAD_REQUEST;
+        ResponseEntity<IUDAnswer> responseEntity = new ResponseEntity<IUDAnswer>(iudAnswer, status);
+        return responseEntity;
     }
 
     @CacheRemoveAll(cacheName = "maintenanceList")
     @RequestMapping(value = "/{id}", method = RequestMethod.DELETE, produces = {MediaType.APPLICATION_JSON_UTF8_VALUE})
-    public ResponseEntity<IUDAnswer> deleteMaintenance(@PathVariable("id") Integer id) {
-        IUDAnswer result = maintenanceService.deleteMaintenance(id);
-
-        HttpStatus status = result.isSuccessful() ? HttpStatus.ACCEPTED : HttpStatus.NOT_FOUND;
-
-        return new ResponseEntity<IUDAnswer>(result, status);
+    public ResponseEntity<IUDAnswer> deleteMaintenance(@PathVariable("id") Integer id, HttpServletRequest request) {
+        User user = (User) request.getSession().getAttribute("USER");
+        IUDAnswer iudAnswer = serviceUtils.checkDeleteForAdmin(user, id);
+        if (!iudAnswer.isSuccessful()) {
+            return new ResponseEntity<IUDAnswer>(iudAnswer, HttpStatus.BAD_REQUEST);
+        }
+        try {
+            iudAnswer = maintenanceService.deleteMaintenance(id);
+        } catch (ReferenceBreakException e) {
+            LOGGER.warn(ON_ENTITY_REFERENCE, e);
+            return new ResponseEntity<IUDAnswer>(new IUDAnswer(id,false, ON_ENTITY_REFERENCE, e.getMessage()), HttpStatus.BAD_REQUEST);
+        } catch (DeletedObjectNotExistsException e) {
+            LOGGER.warn(DELETED_OBJECT_NOT_EXISTS, e);
+            return new ResponseEntity<IUDAnswer>(new IUDAnswer(id, false, DELETED_OBJECT_NOT_EXISTS, e.getMessage()), HttpStatus.BAD_REQUEST);
+        } catch (WrongEntityIdException e) {
+            LOGGER.warn(WRONG_DELETED_ID, e);
+            return new ResponseEntity<IUDAnswer>(new IUDAnswer(id, false, WRONG_DELETED_ID, e.getMessage()), HttpStatus.BAD_REQUEST);
+        } catch (IllegalArgumentException e) {
+            LOGGER.warn(NULL_ID, e);
+            return new ResponseEntity<IUDAnswer>(new IUDAnswer(id, false, NULL_ID, e.getMessage()), HttpStatus.BAD_REQUEST);
+        }
+        HttpStatus status;
+        if (iudAnswer != null && iudAnswer.isSuccessful()) {
+            status = HttpStatus.OK;
+        } else {
+            status = HttpStatus.NOT_ACCEPTABLE;
+        }
+        return new ResponseEntity<IUDAnswer>(iudAnswer, status);
     }
 }
